Split the Variables file into two.
